+++ src/north/north-console/north-console.spec.js
@@ -17,6 +17,7 @@ jest.mock('../../service/certificate.service')
 jest.mock('../../service/encryption.service', () => ({ getInstance: () => ({ decryptText: (password) => password }) }))
 jest.mock('../../service/cache/value-cache.service')
 jest.mock('../../service/cache/file-cache.service')
+jest.mock('../../service/cache/archive.service')
 
 const nowDateString = '2020-02-02T02:02:02.222Z'
 let configuration = null
