+++ src/north/north-file-writer/north-file-writer.spec.js
@@ -14,6 +14,7 @@ jest.mock('../../service/certificate.service')
 jest.mock('../../service/encryption.service', () => ({ getInstance: () => ({ decryptText: (password) => password }) }))
 jest.mock('../../service/cache/value-cache.service')
 jest.mock('../../service/cache/file-cache.service')
+jest.mock('../../service/cache/archive.service')
 
 const nowDateString = '2020-02-02T02:02:02.222Z'
 let configuration = null
