+++ src/north/north-mqtt/north-mqtt.spec.js
@@ -15,6 +15,7 @@ jest.mock('../../service/status.service')
 jest.mock('../../service/encryption.service', () => ({ getInstance: () => ({ decryptText: (password) => password }) }))
 jest.mock('../../service/cache/value-cache.service')
 jest.mock('../../service/cache/file-cache.service')
+jest.mock('../../service/cache/archive.service')
 
 // Mock certificate service
 const CertificateService = jest.mock('../../service/certificate.service')
